face_detection/requirements.txt
onnx
numpy==1.26
scikit-image
scikit-learn
ultralytics
onnx_tool
pillow==10.4.0
pip-system-certs
ffmpeg-python
torch
transformers
onnxruntime
pycocotools
clip
shapely
pyclipper
SentencePiece
imageio
openai-clip
diffusers
cn-clip